+++ src/js_utils/facades.rs
@@ -1,5 +1,4 @@
-use crate::js_utils::adapters::{JsContextAdapter, JsRuntimeAdapter};
-use crate::js_utils::facades::JsValueType::*;
+use crate::js_utils::adapters::JsRuntimeAdapter;
 use crate::js_utils::{JsError, Script};
 use std::future::Future;
 use std::pin::Pin;
@@ -14,11 +13,11 @@ pub trait JsRuntimeBuilder {
 /// handles the logic for transferring data from and to the JsRuntimeAdapter
 pub trait JsRuntimeFacade {
     type JsRuntimeAdapterType: JsRuntimeAdapter;
-    type JsContextFacadeType: JsContextFacade;
 
-    fn js_create_context(&mut self, name: &str) -> Result<(), &str>;
-    fn js_get_main_context(&self) -> &Self::JsContextFacadeType;
-    fn js_get_context(&self, name: &str) -> Option<&Self::JsContextFacadeType>;
+    fn js_realm_create(&mut self, name: &str) -> Result<(), JsError>;
+    fn js_realm_destroy(&mut self, name: &str) -> Result<(), JsError>;
+    fn js_realm_has(&mut self, name: &str) -> Result<bool, JsError>;
+
     fn js_loop_sync<
         R: Send + 'static,
         C: FnOnce(&Self::JsRuntimeAdapterType) -> R + Send + 'static,
@@ -31,46 +30,93 @@ pub trait JsRuntimeFacade {
         consumer: C,
     ) -> Box<dyn Future<Output = R>>;
     fn js_loop_void<C: FnOnce(&Self::JsRuntimeAdapterType) + Send + 'static>(&self, consumer: C);
-}
 
-pub trait JsContextFacade {
-    type JsRuntimeFacadeType: JsRuntimeFacade;
-    type JsContextAdapterType: JsContextAdapter;
+    // realm jobs
 
-    fn js_install_proxy(&self, js_proxy: JsProxy);
+    fn js_loop_realm_sync<
+        R: Send + 'static,
+        C: FnOnce(
+                &Self::JsRuntimeAdapterType,
+                &<<Self as JsRuntimeFacade>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType,
+            ) -> R + Send
+            + 'static ,
+    >(
+        &self,
+        realm_name: Option<&str>,
+        consumer: C,
+    ) -> R;
+    fn js_loop_realm<
+        R: Send + 'static,
+        C: FnOnce(
+                &Self::JsRuntimeAdapterType,
+                &<<Self as JsRuntimeFacade>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType,
+            ) -> Box<dyn Future<Output = R>> + Send
+            + 'static,
+    >(
+        &self,
+        realm_name: Option<&str>,
+        consumer: C,
+    ) -> Box<dyn Future<Output = R>>;
+    fn js_loop_realm_void<
+        C: FnOnce(
+                &Self::JsRuntimeAdapterType,
+                &<<Self as JsRuntimeFacade>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType,
+            ) + Send
+            + 'static,
+    >(
+        &self,
+        realm_name: Option<&str>,
+        consumer: C,
+    );
 
     /// eval a script, please note that eval should not be used for production code, you should always
     /// use modules or functions and invoke them
     /// eval will always need to parse script and some engines like StarLight even require a different syntax (return(1); vs (1);)
+    /// If None is passed as realm_name the default Realm wil be used
     #[allow(clippy::type_complexity)]
     fn js_eval(
         &self,
+        realm_name: Option<&str>,
         script: Script,
     ) -> Pin<Box<dyn Future<Output = Result<Box<dyn JsValueFacade>, JsError>>>>;
 
     // function methods
+    /// Invoke a function and block until the function is done
+    /// If None is passed as realm_name the default Realm wil be used
     fn js_function_invoke_sync(
         &self,
+        realm_name: Option<&str>,
         namespace: &[&str],
         method_name: &str,
         args: Vec<Box<dyn JsValueFacade>>,
     ) -> Result<Box<dyn JsValueFacade>, JsError>;
 
-    fn js_loop_sync<
-        R : Send + 'static,
-        C: FnOnce(&<<Self as JsContextFacade>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType, &Self::JsContextAdapterType) -> R + Send + 'static,
-    >(
+    /// Invoke a function
+    /// this returns a Future which will fulfill when the function is done
+    /// If None is passed as realm_name the default Realm wil be used
+    #[allow(clippy::type_complexity)]
+    fn js_function_invoke(
         &self,
-        consumer: C,
-    ) -> R;
-    fn js_loop<R : Send + 'static, C>(&self, consumer: C) -> Pin<Box<dyn Future<Output = Result<R, JsError>>>>
-    where
-        C: FnOnce(&<<Self as JsContextFacade>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType, &Self::JsContextAdapterType) -> Result<R, JsError> + Send + 'static;
-    fn js_loop_void<C>(&self, consumer: C) where C: FnOnce(&<<Self as JsContextFacade>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType, &Self::JsContextAdapterType) + Send + 'static;
+        realm_name: Option<&str>,
+        namespace: &[&str],
+        method_name: &str,
+        args: Vec<Box<dyn JsValueFacade>>,
+    ) -> Pin<Box<dyn Future<Output = Result<Box<dyn JsValueFacade>, JsError>>>>;
+
+    /// Invoke a function without waiting for a result
+    /// If None is passed as realm_name the default Realm wil be used
+    fn js_function_invoke_void(
+        &self,
+        realm_name: Option<&str>,
+        namespace: &[&str],
+        method_name: &str,
+        args: Vec<Box<dyn JsValueFacade>>,
+    );
 }
 
-/// The JsValueFacade is a thread safe Sendable representation of a variable in the Script engine
+/// The JsValueFacade is a Send-able representation of a variable in the Script engine
 
+#[derive(PartialEq)]
 pub enum JsValueType {
     I32,
     F64,
@@ -83,26 +129,21 @@ pub enum JsValueType {
     Date,
     Null,
     Undefined,
+    Array,
 }
 
 pub trait JsValueFacade: Send + Sync {
-    fn js_is_i32(&self) -> bool {
-        false
-    }
-    fn js_is_bool(&self) -> bool {
-        false
-    }
-    fn js_is_null(&self) -> bool {
-        false
-    }
-    fn js_is_undefined(&self) -> bool {
-        false
-    }
     fn js_is_null_or_undefined(&self) -> bool {
-        self.js_is_null() || self.js_is_undefined()
+        self.js_get_type() == JsValueType::Null || self.js_get_type() == JsValueType::Undefined
     }
     fn js_as_i32(&self) -> i32 {
         panic!("not an i32");
+    }
+    fn js_as_f64(&self) -> f64 {
+        panic!("not an f64");
+    }
+    fn js_as_str(&self) -> &str {
+        panic!("not a String");
     }
     fn js_as_bool(&self) -> bool {
         panic!("not a bool");
@@ -114,48 +155,53 @@ pub struct JsNull {}
 pub struct JsUndefined {}
 
 impl JsValueFacade for JsNull {
-    fn js_is_null(&self) -> bool {
-        true
-    }
-
     fn js_get_type(&self) -> JsValueType {
-        Null
+        JsValueType::Null
     }
 }
 
 impl JsValueFacade for JsUndefined {
-    fn js_is_undefined(&self) -> bool {
-        true
-    }
     fn js_get_type(&self) -> JsValueType {
-        Undefined
+        JsValueType::Undefined
     }
 }
 
 impl JsValueFacade for i32 {
-    fn js_is_i32(&self) -> bool {
-        true
-    }
-
     fn js_as_i32(&self) -> i32 {
         *self
     }
 
     fn js_get_type(&self) -> JsValueType {
-        I32
+        JsValueType::I32
+    }
+}
+
+impl JsValueFacade for f64 {
+    fn js_as_f64(&self) -> f64 {
+        *self
+    }
+
+    fn js_get_type(&self) -> JsValueType {
+        JsValueType::F64
     }
 }
 
 impl JsValueFacade for bool {
-    fn js_is_bool(&self) -> bool {
-        true
-    }
-
     fn js_as_bool(&self) -> bool {
         *self
     }
 
     fn js_get_type(&self) -> JsValueType {
-        Boolean
+        JsValueType::Boolean
+    }
+}
+
+impl JsValueFacade for String {
+    fn js_as_str(&self) -> &str {
+        self.as_str()
+    }
+
+    fn js_get_type(&self) -> JsValueType {
+        JsValueType::String
     }
 }